Type the connection-budget fraction so it survives the wire
The fraction of an instance's headroom we are willing to consume sat as an
untyped parameter inside integer arithmetic, so Postgres resolved the
multiplication as integer times integer and typed the parameter an integer.
0.15 arrived as 0. Every instance fell back to its floor as its budget, and
any job weighing more than the floor stopped being claimable at all.

Nothing failed and nothing was logged. Observation weighs one, so it still
fit under the floor and carried on as normal; migrations weigh up to eight,
so they sat pending forever with no error to explain it. Running the same
statement by hand could not reproduce it, because writing the statement out
declares the parameter's type.

The test claims work heavier than the floor against known headroom, which is
the only shape of test that would have caught this.

+++ internal/store/store.go
@@ -442,7 +442,13 @@ func (s *Store) ClaimJob(ctx context.Context, workerID string, lease time.Durati
 		            <= GREATEST($3, LEAST($4, FLOOR((
 		                   COALESCE(i.max_connections, 100)
 		                 - COALESCE(i.reserved_connections, 3)
-		                 - COALESCE(i.used_connections, 0)) * $5)::int)))
+		                 -- $5 is cast explicitly. Left to inference, Postgres
+		                 -- resolves "integer * unknown" to integer arithmetic
+		                 -- and types the parameter as an integer, which silently
+		                 -- truncates the fraction to zero: every instance then
+		                 -- gets the floor as its budget, and work heavier than
+		                 -- the floor is never claimable at all.
+		                 - COALESCE(i.used_connections, 0)) * $5::float8)::int)))
 		      ORDER BY j.run_after
 		      -- Only the job row is locked: an outer join's nullable side cannot
 		      -- be, and the instance row is read rather than claimed.
